requirements.txt
PyYAML~=5.3.1
pycurl~=7.43.0.5
ldap3~=2.7
SATOSA~=6.1.0
setuptools~=46.1.3
pytest~=5.4.1